refactor: split crypto helpers from ObliviousX
